+++ CHANGELOG.md
@@ -2,7 +2,9 @@
  - 2018-xx-xx
 
 ### Changes since the last release
+
  * GUI window would reset to screen center on a game restart if it was hidden in the previous game, now fixed.
+ * In-Flight trajectory ground target marker would sometimes appear on screen when behind the camera view, now fixed.
  
 ### For Developers